+++ src/background/services/IpcBackground.ts
@@ -1,7 +1,8 @@
-import {ActionType} from "../shared/enums/ActionType";
+import {ActionType} from "../../shared/enums/ActionType";
 
-export class IpcBackground {
+export abstract class IpcBackground {
     static sendMessage(text: string, responseCallback?: ResponseCallback): void {
+        console.log(`[torrent2box - background]: ${text}`);
         chrome.tabs.query({active: true}, (result) => {
             chrome.tabs.sendMessage(result[0].id!, {
                 actionType: ActionType.Message,