+++ src/background/utilities/fetch-instance.ts
@@ -1,0 +1,3 @@
+import {UnifiedFetch} from "unified-fetch";
+
+export const fetchNoHooks = new UnifiedFetch();