+++ src/options/interfaces/Form.ts
@@ -1,0 +1,10 @@
+import {ValidationErrors} from "meta-validator";
+
+export interface Form {
+    formModel: any;
+    validationErrors: ValidationErrors;
+    isComplete: boolean;
+    isSkipMissingProperties: boolean;
+    isFormValid(): Promise<boolean>;
+    resetForm(): void;
+}