+++ src/shared/enums/ActionType.ts
@@ -1,5 +1,4 @@
 export enum ActionType {
     Message = "Message",
-    Notify = "Notify",
     AddTorrent = "AddTorrent"
 }